chore: removes extra check examples
