--- src/checkbox/examples-list.js
@@ -1,22 +1,0 @@
-/*
-Copyright (c) 2018 Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-
-/* eslint-env node */
-/* eslint-disable flowtype/require-valid-file-annotation */
-module.exports = {
-  SIMPLE_EXAMPLE: 'Checkbox example',
-  MULTILINE: 'Multiline checkbox',
-  WITH_ERROR: 'Checkbox with an error',
-  INDETERMINATE: 'Checkbox Indeterminate',
-  DISABLED: 'Checkbox disabled',
-  LEFT_ALIGN: 'Checkbox label placed left',
-  BOTTOM_ALIGN: 'Checkbox label placed bottom',
-  STYLES_OVERRIDES: 'Checkbox with style overrides',
-  COMPONENTS_OVERRIDES: 'Checkbox with override components',
-  EXTRA_PROPS: 'Checkbox with extra props',
-  WITH_FOCUS: 'Checkbox with a ref to focus on',
-};
